Remove WLAN Eth from config file. We obtain it from the interface. Continuous screenshot is called using Popen

+++ scrape/channel_surfer.py
@@ -18,6 +18,7 @@ from os.path import join
 from shlex import split
 import wave
 import pyaudio
+import fcntl, socket, struct
 
 LOG_DIR = os.getenv("LogDir")
 LOG_FILE = 'channel_surfer.log'
@@ -32,6 +33,12 @@ RATE = 44100
 
 PLAT = os.getenv("PLATFORM")
 PLATFORM_DIR = os.getenv("PLATFORM_DIR")
+
+
+def getHwAddr(ifname):
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    info = fcntl.ioctl(s.fileno(), 0x8927,  struct.pack('256s', bytes(ifname[:15], 'utf-8')))
+    return ''.join(['%02x:' % b for b in info[18:24]])[:-1]
 
 class SurferAborted(Exception):
     """Raised when we encounter an error while surfing this channel."""
@@ -62,7 +69,7 @@ class ChannelSurfer(object):
         # On Roku: Start a background process that continuously captures screenshots to
         # the same file: ${LogDir}/continuous_screenshot.png
         if PLAT == "ROKU":
-            subprocess.call(join(PLATFORM_DIR, 'scripts') + '/capture_screenshot.sh', shell=True)
+            subprocess.Popen(join(PLATFORM_DIR, 'scripts') + '/capture_screenshot.sh', shell=True)
 
     def log(self, *args):
 
@@ -189,15 +196,20 @@ class ChannelSurfer(object):
         )
         #eth_mac = os.environ['ETH_MAC_ADDRESS']
         #lan_if_name = os.environ['LANIF']
-        wlan_eth_mac = os.environ['WLAN_ETH_MAC_ADDRESS']
+
         wlan_if_name = os.environ['WLANIF']
+        wlan_eth_mac = getHwAddr(wlan_if_name)
 
         pcap_path = join(str(self.pcap_dir), str(self.pcap_filename))
         # tcpdump -v -w "$1".pcap -i ${LANIF} ether host $ETH_MAC_ADDRESS and not arp and port not ssh
-        self.tcpdump_proc = subprocess.Popen(
-            ['tcpdump', '-w', pcap_path, '-i', wlan_if_name, 'ether',  'host',
-            wlan_eth_mac] + split('and not arp and port not ssh'),
-            stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+
+        command = ['tcpdump', '-w', pcap_path, '-i', wlan_if_name, 'ether',  'host',
+                   wlan_eth_mac] + split(' and not arp and port not ssh')
+        self.tcpdump_proc = subprocess.Popen(command,
+            stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+
+        self.log("tcpdump command executed: "+ " ".join(command))
+        self.log("tcpdump >> " + self.tcpdump_proc.stdout.readline().decode("utf-8").rstrip())
 
         #self.log('./start_pcap.sh ' + str(self.pcap_dir) + "/" + str(self.pcap_filename))
         #subprocess.Popen(
@@ -209,6 +221,7 @@ class ChannelSurfer(object):
     def capture_screenshots(self, timeout):
 
         start_time = time.time()
+        err_reported = False
         FFMPEG_SCREENSHOT_NAME = os.path.join(LOG_DIR, 'continuous_screenshot.png')
         while time.time() - start_time <= timeout:
             t0 = time.time()
@@ -220,7 +233,12 @@ class ChannelSurfer(object):
             # capture_screenshot.sh continuously writes the latest screenshot
             # images to continuous_screenshot
             if PLAT == "ROKU":
-                copy2(FFMPEG_SCREENSHOT_NAME, screenshot_filename)
+                if os.path.exists(FFMPEG_SCREENSHOT_NAME):
+                    copy2(FFMPEG_SCREENSHOT_NAME, screenshot_filename)
+                else:
+                    if not err_reported:
+                        self.log("Error! Screenshot file %s is missing." % FFMPEG_SCREENSHOT_NAME)
+                        err_reported = True
             elif PLAT == "AMAZON":
                 self.rrc.take_screenshot(screenshot_filename)
